courseinfo part 2 step 10
